+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/AutoOpMode/AutoOpMode.java
@@ -2,12 +2,24 @@ package org.firstinspires.ftc.teamcode.AutoOpMode;
 import static com.qualcomm.robotcore.util.ElapsedTime.Resolution.MILLISECONDS;
 
 import com.acmerobotics.roadrunner.geometry.Pose2d;
+import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.util.ElapsedTime;
+
+import org.firstinspires.ftc.robotcore.external.ClassFactory;
+import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
+import org.firstinspires.ftc.robotcore.external.navigation.VuforiaLocalizer;
+import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackable;
+import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackableDefaultListener;
+import org.firstinspires.ftc.robotcore.external.tfod.Recognition;
+import org.firstinspires.ftc.robotcore.external.tfod.TFObjectDetector;
 import org.firstinspires.ftc.teamcode.Subsystems.DriveTrain;
 
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
+
+import java.util.List;
 
 /**
  * Write a detailed description of the autoClass
@@ -15,25 +27,283 @@ import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 //TODO: Copy and Rename Autonomous Mode
 @Autonomous(name = "Autonomous World 1", group = "00-Autonomous" , preselectTeleOp = "TeleOp")
 public class AutoOpMode extends LinearOpMode{
+    public static final Vector2d ORIGIN = new Vector2d(0,0);
+    public static final Pose2d ORIGINPOSE = new Pose2d(0,0,Math.toRadians(0));
+
+    // Declare and assign starting pose of robot
+    public static final Pose2d STARTPOS_1 =  new Pose2d(-61,7,Math.toRadians(180));
+    public static final Pose2d STARTPOS_2 =  new Pose2d(-61,-40,Math.toRadians(180));
+    public static final Pose2d STARTPOS_3 =  new Pose2d(61,7,Math.toRadians(0));
+    public static final Pose2d STARTPOS_4 =  new Pose2d(61,-40,Math.toRadians(0));
+
+
+    //Define and declare Playing Alliance
+    public enum PLAYING_ALLIANCE{
+        RED_ALLIANCE,
+        BLUE_ALLIANCE,
+    }
+    public static PLAYING_ALLIANCE playingAlliance = PLAYING_ALLIANCE.BLUE_ALLIANCE;
+    public static double ALLIANCE_FACTOR = 1;
+
+
+    //Define and declare Robot Starting Locations
+    public enum START_POSITION{
+        POS1,
+        POS2,
+        POS3,
+        POS4
+    }
+
+    //Set a default start and park position
+    public static START_POSITION startPosition;
+    public static PARKING_LOCATION parkingLocation = PARKING_LOCATION.PARKPOS_1;
+
+    //Define and declare Robot parking locations
+    public enum PARKING_LOCATION{
+        PARKPOS_1,
+        PARKPOS_2,
+        PARKPOS_3
+    }
+
+
+    //Select the different colors on the signal sleeve, and define them here
+    public enum VISION_IDENTIFIED_TARGET {
+        LOCATION_1,
+        LOCATION_2,
+        LOCATION_3,
+    };
+
+    public enum VISION_IDENTIFIER{
+        BOLT,
+        BULB,
+        PANEL,
+        GREEN,
+        YELLOW,
+        PURPLE
+    };
+    public static VISION_IDENTIFIER visionIdentifier = VISION_IDENTIFIER.BOLT;
+
+    //Static fields to pass Pos from Autonomous to TeleOp
+    public static boolean poseSetInAutonomous = false;
+    public static Pose2d currentPose = new Pose2d();
+
+    public enum VISION_STATE {
+        TFOD_INIT,
+        TFOD_ACTIVE,
+        TFOD_RUNNING,
+        INACTIVE
+    }
+    public VISION_STATE visionState = VISION_STATE.INACTIVE;
+
+    public enum ACTIVE_WEBCAM{
+        WEBCAM1,
+    }
+
+
+    /*
+     * IMPORTANT: You need to obtain your own license key to use Vuforia. The string below with which
+     * 'parameters.vuforiaLicenseKey' is initialized is for illustration only, and will not function.
+     * A Vuforia 'Development' license key, can be obtained free of charge from the Vuforia developer
+     * web site at https://developer.vuforia.com/license-manager.
+     *
+     * Vuforia license keys are always 380 characters long, and look as if they contain mostly
+     * random data. As an example, here is a example of a fragment of a valid key:
+     *      ... yIgIzTqZ4mWjk9wd3cZO9T1axEqzuhxoGlfOOI2dRzKS4T0hQ8kT ...
+     * Once you've obtained a license key, copy the string from the Vuforia web site
+     * and paste it in to your code on the next line, between the double quotes.
+     */
+    private static final String VUFORIA_KEY =
+            "AZME4Mr/////AAABmY+MAyxxT0IileR7JBqaAPsxN2XNYlaGBtEjYaHOlVVTqPQf7NH9eIrosYKKQHPGEXLtJUsdMwZ9e3EXBfy6arulcLPvdpW9bqAB2F2MJJXo35lLA096l/t/LQTi+etVso0Xc5RYkTVSIP3YABp1TeOaF8lCSpjVhPIVW3l/c/XlrnEMPhJk9IgqMEp4P/ifqAqMMMUAIKPEqIrXIv79TvAfdIJig46gfQGaQl5tFHr3nmvMbh/LhFrh5AWAy3B/93cCkOszmYkdHxZStbNB5lMdkTnf3sCnYbQY4jviorfhYrAkqHWH6vNOB9lUt8dOSeHsDtlk33e/6xQgOCNYFN80anYMp82JNDBFX3oyGliV";
+
+
+    // Class Members
+    //private OpenGLMatrix lastLocation = null;
+    private VuforiaLocalizer vuforia = null;
+    //private VuforiaTrackables targets   = null ;
+
+    /**
+     * This is the webcam we are to use. As with other hardware devices such as motors and
+     * servos, this device is identified using the robot configuration tool in the FTC application.
+     */
+    WebcamName webcamName = null;
+
+    public  boolean targetVisible = false;
+    private float phoneXRotate    = 0;
+    private float phoneYRotate    = 0;
+    private float phoneZRotate    = 0;
+
+    public VuforiaLocalizer.Parameters parameters;
+
+    public List<VuforiaTrackable> allTrackables;
+
+    //Tensor Flow parameters
+    /* Note: This sample uses the all-objects Tensor Flow model (PowerPlay.tflite), which contains
+     * the following 3 detectable objects
+     *  0: Bolt,
+     *  1: Bulb,
+     *  2: Panel,
+     */
+    private static final String TFOD_MODEL_ASSET = "PowerPlay.tflite";
+    // private static final String TFOD_MODEL_FILE  = "/sdcard/FIRST/tflitemodels/CustomTeamModel.tflite";
+
+
+    private static final String[] LABELS = {
+            "1 Bolt",
+            "2 Bulb",
+            "3 Panel",
+            "4 Green",
+            "5 Yellow",
+            "6 Purple"
+    };
+
+    public String detectedLabel = "None";
+
+
+    private TFObjectDetector tfod;
+    private List<Recognition> recognitions;
+    public VISION_IDENTIFIED_TARGET targetLevelDetected = VISION_IDENTIFIED_TARGET.LOCATION_1;
 
         public boolean DEBUG_FLAG = true;
 
         public DriveTrain driveTrain;
-        public Vision vision;
-
-
-        //need to fix vision parking positions
-        public Vision.PARKING_LOCATION parkingLocation = Vision.PARKING_LOCATION.PARKPOS_1;
 
         boolean parked = false ;
         public ElapsedTime gameTimer = new ElapsedTime(MILLISECONDS);;
 
-        public Vision.ACTIVE_WEBCAM activeWebcam = Vision.ACTIVE_WEBCAM.WEBCAM1;
-        public Vision.VISION_IDENTIFIED_TARGET targetZone = Vision.VISION_IDENTIFIED_TARGET.LOCATION_1;//Set a default vision value
+        public ACTIVE_WEBCAM activeWebcam = ACTIVE_WEBCAM.WEBCAM1;
+        public VISION_IDENTIFIED_TARGET targetZone = VISION_IDENTIFIED_TARGET.LOCATION_1;//Set a default vision value
         int targetZoneLevel = 0;
 
-        double af = Vision.ALLIANCE_FACTOR;
+        double af = ALLIANCE_FACTOR;
 
+    /**
+     * Initialize the Vuforia localization engine.
+     */
+    public AutoOpMode(HardwareMap hardwareMap, ACTIVE_WEBCAM activeWebcam) {
+        activeWebcam = ACTIVE_WEBCAM.WEBCAM1;
+
+        if (activeWebcam == ACTIVE_WEBCAM.WEBCAM1){
+            webcamName = hardwareMap.get(WebcamName.class, "Webcam1");
+        } /*else { //TODO: Uncomment if using 2 cameras;
+            webcamName = hardwareMap.get(WebcamName.class, "Webcam2");
+        }
+        */
+
+        /*
+         * Configure Vuforia by creating a Parameter object, and passing it to the Vuforia engine.
+         * We can pass Vuforia the handle to a camera preview resource (on the RC phone);
+         * If no camera monitor is desired, use the parameter-less constructor instead (commented out below).
+         */
+        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+        parameters = new VuforiaLocalizer.Parameters(cameraMonitorViewId);
+
+        parameters.vuforiaLicenseKey = VUFORIA_KEY;
+
+        /**
+         * We also indicate which camera on the RC we wish to use.
+         */
+        parameters.cameraName = webcamName;
+
+        // Make sure extended tracking is disabled for this example.
+        parameters.useExtendedTracking = false;
+
+        //  Instantiate the Vuforia engine
+        vuforia = ClassFactory.getInstance().createVuforia(parameters);
+        initTfod(hardwareMap);
+    }
+
+    /**
+     * Initialize the TensorFlow Object Detection engine.
+     */
+    private void initTfod(HardwareMap hardwareMap) {
+        int tfodMonitorViewId = hardwareMap.appContext.getResources().getIdentifier(
+                "tfodMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+        TFObjectDetector.Parameters tfodParameters = new TFObjectDetector.Parameters(tfodMonitorViewId);
+        tfodParameters.minResultConfidence = 0.8f;
+        tfodParameters.isModelTensorFlow2 = true;
+        tfodParameters.inputSize = 320; //320;
+        tfod = ClassFactory.getInstance().createTFObjectDetector(tfodParameters, vuforia);
+        tfod.loadModelFromAsset(TFOD_MODEL_ASSET, LABELS);
+
+        visionState = VISION_STATE.TFOD_INIT;
+
+    }
+    /**
+     * Activate Vuforia Tensor Flow to determine target zone
+     * This is to be done at Init in Autonomous mode
+     */
+    public void activateVuforiaTensorFlow(){
+        /**
+         * Activate TensorFlow Object Detection before we wait for the start command.
+         * Do it here so that the Camera Stream window will have the TensorFlow annotations visible.
+         **/
+        if (tfod != null) {
+            tfod.activate();
+            visionState = VISION_STATE.TFOD_ACTIVE;
+
+            // The TensorFlow software will scale the input images from the camera to a lower resolution.
+            // This can result in lower detection accuracy at longer distances (> 55cm or 22").
+            // If your target is at distance greater than 50 cm (20") you can adjust the magnification value
+            // to artificially zoom in to the center of image.  For best results, the "aspectRatio" argument
+            // should be set to the value of the images used to create the TensorFlow Object Detection model
+            // (typically 1.78 or 16/9).
+
+            // Uncomment the following line if you want to adjust the magnification and/or the aspect ratio of the input images.
+            //tfod.setZoom(1.75, 16.0/9.0);
+            tfod.setZoom(1.0, 16.0/9.0);
+            recognitions = tfod.getUpdatedRecognitions();
+        }
+    }
+
+    /**
+     * Run Tensor flow algorithm.
+     * This is to be run till the play button is pressed.. the last target zone identified is returned.
+     * @return
+     */
+
+
+
+    public VISION_IDENTIFIED_TARGET runVuforiaTensorFlow() {
+        visionState = VISION_STATE.TFOD_RUNNING;
+        if (tfod != null) {
+            // getUpdatedRecognitions() will return null if no new information is available since
+            // the last time that call was made.
+            recognitions = tfod.getUpdatedRecognitions();
+            if (recognitions != null) {
+                //telemetry.addData("# Object Detected", updatedRecognitions.size());
+                for (Recognition recognition : recognitions) {
+                    // check label to see which target zone to go after.
+                    detectedLabel = recognition.getLabel();
+                    switch(detectedLabel){
+                        case "1 Bolt":
+                        case "6 Purple":
+                            targetLevelDetected = VISION_IDENTIFIED_TARGET.LOCATION_1;
+                            break;
+                        case "2 Bulb":
+                        case "5 Yellow":
+                            targetLevelDetected = VISION_IDENTIFIED_TARGET.LOCATION_2;
+                            break;
+                        case "3 Panel":
+                        case "4 Green":
+                            targetLevelDetected = VISION_IDENTIFIED_TARGET.LOCATION_3;
+                            break;
+                    }
+                }
+            }
+        }
+
+        return targetLevelDetected;
+    }
+
+    /**
+     * Stop Tensor Flow algorithm
+     */
+    public void deactivateVuforiaTensorFlow(){
+        if (tfod != null) {
+            tfod.shutdown();
+            visionState = VISION_STATE.INACTIVE;
+        }
+    }
         @Override
         public void runOpMode() throws InterruptedException {
             /*Create your Subsystem Objects*/
@@ -44,11 +314,10 @@ public class AutoOpMode extends LinearOpMode{
 
             //Key Pay inputs to select Game Plan;
             selectGamePlan();
-            vision = new Vision(hardwareMap, activeWebcam);
-            af = Vision.ALLIANCE_FACTOR;
+            af = ALLIANCE_FACTOR;
 
             // Initiate Camera on Init.
-            vision.activateVuforiaTensorFlow();
+            activateVuforiaTensorFlow();
 
             buildAuto();
 
@@ -62,7 +331,7 @@ public class AutoOpMode extends LinearOpMode{
             while (!isStopRequested()) {
 
                 //Run Vuforia Tensor Flow
-                targetZone = vision.runVuforiaTensorFlow();
+                targetZone = runVuforiaTensorFlow();
                 targetZoneLevel = targetZone.ordinal();
 
                 if (!parked) {
@@ -78,7 +347,7 @@ public class AutoOpMode extends LinearOpMode{
                 while (opModeIsActive() && !isStopRequested() && !parked) {
                     gameTimer.reset();
 
-                    vision.deactivateVuforiaTensorFlow();
+                    deactivateVuforiaTensorFlow();
 
 
 
@@ -90,8 +359,8 @@ public class AutoOpMode extends LinearOpMode{
 
                     //Write last position to static class to be used as initial position in TeleOp
 
-                    vision.currentPose = driveTrain.getPoseEstimate();
-                    vision.poseSetInAutonomous = true;
+                    currentPose = driveTrain.getPoseEstimate();
+                    poseSetInAutonomous = true;
 
 
                     if (DEBUG_FLAG) {
@@ -102,8 +371,8 @@ public class AutoOpMode extends LinearOpMode{
 
             }
             safeWait(100);
-            vision.currentPose = driveTrain.getPoseEstimate();
-            vision.poseSetInAutonomous = true;
+            currentPose = driveTrain.getPoseEstimate();
+            poseSetInAutonomous = true;
         }
 
         //Initialize any other TrajectorySequences as desired
@@ -117,15 +386,15 @@ public class AutoOpMode extends LinearOpMode{
         Pose2d parkPose; //4 different poses
 
         public void buildAuto(){
-            if(vision.playingAlliance == Vision.PLAYING_ALLIANCE.BLUE_ALLIANCE){
-                initPose = Vision.STARTPOS_1; //Starting pos when on blue alliance
+            if(playingAlliance == PLAYING_ALLIANCE.BLUE_ALLIANCE){
+                initPose = currentPose; //Starting pos when on blue alliance
                 midWayPose = new Pose2d(0,0,Math.toRadians(0)); //Choose the pose to move forward towards signal cone
                 pickConePose =  new Pose2d(0,0,Math.toRadians(0)); //Choose the pose to move to the stack of cones
                 dropCone =  new Pose2d(0,0,Math.toRadians(0)); //Choose the pose to move to the stack of cones
                 parkPose =  new Pose2d(0,0,Math.toRadians(0)); //Choose the pose to move to the stack of cones
 
             } else {
-                initPose = Vision.STARTPOS_2; //Starting pos when on red alliance
+                initPose = currentPose; //Starting pos when on red alliance
                 midWayPose = new Pose2d(0,0,Math.toRadians(0)); //Choose the pose to move forward towards cone
                 pickConePose =  new Pose2d(0,0,Math.toRadians(0)); //Choose the pose to move to the stack of cones
                 dropCone =  new Pose2d(0,0,Math.toRadians(0)); //Choose the pose to move to the stack of cones
@@ -136,52 +405,55 @@ public class AutoOpMode extends LinearOpMode{
             //Pick 5 cones and park
             trajInitToPickAndDropConeToPark = driveTrain.trajectorySequenceBuilder(initPose)
                     .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(pickConePose)
                     .addTemporalMarker(0,()->{
                         pickCone();
                     })
-                    .lineToLinearHeading(pickConePose)
-                    .addTemporalMarker(0,() -> {
-                        turnToPole();
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(dropCone)
+                    .addTemporalMarker(0,()->{
                         dropCone();
                     })
-                    .lineToLinearHeading(dropCone)
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(pickConePose)
                     .addTemporalMarker(0,()->{
                         pickCone();
                     })
-                    .lineToLinearHeading(pickConePose)
-                    .addTemporalMarker(0,() -> {
-                        turnToPole();
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(dropCone)
+                    .addTemporalMarker(0,()->{
                         dropCone();
                     })
-                    .lineToLinearHeading(dropCone)
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(pickConePose)
                     .addTemporalMarker(0,()->{
                         pickCone();
                     })
-                    .lineToLinearHeading(pickConePose)
-                    .addTemporalMarker(0,() -> {
-                        turnToPole();
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(dropCone)
+                    .addTemporalMarker(0,()->{
                         dropCone();
                     })
-                    .lineToLinearHeading(dropCone)
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(pickConePose)
                     .addTemporalMarker(0,()->{
                         pickCone();
                     })
-                    .lineToLinearHeading(pickConePose)
-                    .addTemporalMarker(0,() -> {
-                        turnToPole();
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(dropCone)
+                    .addTemporalMarker(0,()->{
                         dropCone();
                     })
-                    .lineToLinearHeading(dropCone)
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(pickConePose)
                     .addTemporalMarker(0,()->{
                         pickCone();
                     })
-                    .lineToLinearHeading(pickConePose)
-                    .addTemporalMarker(0,() -> {
-                        turnToPole();
+                    .lineToLinearHeading(midWayPose)
+                    .lineToLinearHeading(dropCone)
+                    .addTemporalMarker(0,()->{
                         dropCone();
                     })
-                    .lineToLinearHeading(dropCone)
-                    .lineToLinearHeading(parkPose)
                     .build();
         }
 
@@ -193,11 +465,6 @@ public class AutoOpMode extends LinearOpMode{
 
         //Write a method which is able to pick the cone depending on your subsystems
         public void pickCone(){
-
-        }
-
-        //Write a method which is able to turn to the pole depending on your subsystems
-        public void turnToPole(){
 
         }
 
@@ -237,15 +504,15 @@ public class AutoOpMode extends LinearOpMode{
 
             while(!isStopRequested()){
                 if(gamepad1.x){
-                    Vision.playingAlliance = Vision.PLAYING_ALLIANCE.BLUE_ALLIANCE;
-                    Vision.ALLIANCE_FACTOR = -1;
-                    telemetry.addData("Playing Alliance Selected : ", Vision.playingAlliance);
+                    playingAlliance = PLAYING_ALLIANCE.BLUE_ALLIANCE;
+                    ALLIANCE_FACTOR = -1;
+                    telemetry.addData("Playing Alliance Selected : ", playingAlliance);
                     break;
                 }
                 if(gamepad1.b){
-                    Vision.playingAlliance = Vision.PLAYING_ALLIANCE.RED_ALLIANCE;
-                    Vision.ALLIANCE_FACTOR = 1;
-                    telemetry.addData("Playing Alliance Selected : ", Vision.playingAlliance);
+                    playingAlliance = PLAYING_ALLIANCE.RED_ALLIANCE;
+                    ALLIANCE_FACTOR = 1;
+                    telemetry.addData("Playing Alliance Selected : ", playingAlliance);
                     break;
                 }
                 telemetry.update();
@@ -260,32 +527,32 @@ public class AutoOpMode extends LinearOpMode{
                 telemetry.addData("Blue Right: (Y)", "");
                 telemetry.addData("Red Left: (B)", "");
                 telemetry.addData("Red Right: (A)", "");
-                telemetry.addData("Playing Alliance Selected: ", Vision.playingAlliance);
+                telemetry.addData("Playing Alliance Selected: ", playingAlliance);
                 if(gamepad1.x){
-                    Vision.startPosition = Vision.START_POSITION.POS1;
-                    telemetry.addData("Start Position: ", Vision.startPosition);
+                    startPosition = START_POSITION.POS1;
+                    telemetry.addData("Start Position: ", startPosition);
                     break;
                 }
                 if(gamepad1.y){
-                    Vision.startPosition = Vision.START_POSITION.POS2;
-                    telemetry.addData("Start Position: ", Vision.startPosition);
+                    startPosition = START_POSITION.POS2;
+                    telemetry.addData("Start Position: ", startPosition);
                     break;
                 }
                 if(gamepad1.b){
-                    Vision.startPosition = Vision.START_POSITION.POS3;
-                    telemetry.addData("Start Position: ", Vision.startPosition);
+                    startPosition = START_POSITION.POS3;
+                    telemetry.addData("Start Position: ", startPosition);
                     break;
                 }
                 if(gamepad1.a){
-                    Vision.startPosition = Vision.START_POSITION.POS4;
-                    telemetry.addData("Start Position: ", Vision.startPosition);
+                    startPosition = START_POSITION.POS4;
+                    telemetry.addData("Start Position: ", startPosition);
                     break;
                 }
                 telemetry.update();
             }
             telemetry.clearAll();
-            telemetry.addData("Playing Alliance Selected : ", Vision.playingAlliance);
-            telemetry.addData("StartPose : ", Vision.startPosition);
+            telemetry.addData("Playing Alliance Selected : ", playingAlliance);
+            telemetry.addData("StartPose : ", startPosition);
             telemetry.update();
             safeWait(200);
         }
@@ -299,21 +566,21 @@ public class AutoOpMode extends LinearOpMode{
             telemetry.addData("DEBUG_FLAG is : ", DEBUG_FLAG);
 
 
-            telemetry.addData("Playing Alliance Selected : ", vision.playingAlliance);
-            telemetry.addData("Start Position : ", vision.startPosition);
-            telemetry.addData("Parking Location : ", vision.parkingLocation);
+            telemetry.addData("Playing Alliance Selected : ", playingAlliance);
+            telemetry.addData("Start Position : ", startPosition);
+            telemetry.addData("Parking Location : ", parkingLocation);
 
 
-            telemetry.addData("Vision.playingAlliance : ", vision.playingAlliance);
-            telemetry.addData("Vision.poseSetInAutonomous : ", vision.poseSetInAutonomous);
-            telemetry.addData("Vision.currentPose : ", vision.currentPose);
+            telemetry.addData("Vision.playingAlliance : ", playingAlliance);
+            telemetry.addData("Vision.poseSetInAutonomous : ", poseSetInAutonomous);
+            telemetry.addData("Vision.currentPose : ", currentPose);
 
             //****** Drive debug ******
             telemetry.addData("Drive Mode : ", driveTrain.driveMode);
             telemetry.addData("PoseEstimate :", driveTrain.poseEstimate);
             //telemetry.addData("Battery Power", driveTrain.getBatteryVoltage(hardwareMap));
 
-            telemetry.addData("Vision detectedLabel", vision.detectedLabel);
+            telemetry.addData("Vision detectedLabel", detectedLabel);
             telemetry.addData("Vision targetZone :", targetZone);
             telemetry.addData("Vision targetZoneLevel :", targetZoneLevel);
